- load image with api speficication      + baseURL + size + imagePath)      + optimize image url with CGSize -> try to detect suitable `size` before create url
- store default image configuration in local
- fetch configuration from server and update in local
- fix minor UI bugs: cell UI, text label

+++ Challenge/Models/MovieDetail.swift
@@ -13,7 +13,7 @@ struct MovieDetail: MovieType, Decodable {
     let title: String
     let duration: Int?
     let voteAverage: Double
-    let totalVotes: Double
+    let totalVotes: Int
     let posterPath: String?
     let backdropPath: String?
     let overview: String
